+++ packages/components/src/components/MarkdownEditor/MarkdownEditor.tsx
@@ -1,10 +1,4 @@
-import React, {
-  type KeyboardEventHandler,
-  type RefObject,
-  useEffect,
-  useRef,
-  useState,
-} from "react";
+import { type KeyboardEventHandler, useState } from "react";
 import styles from "./MarkdownEditor.module.scss";
 import { Markdown, type MarkdownProps } from "@/components/Markdown";
 import { TextArea, type TextAreaProps } from "@/components/TextArea";
@@ -12,9 +6,8 @@ import { Toolbar } from "@/components/MarkdownEditor/components/Toolbar";
 import clsx from "clsx";
 import { flowComponent } from "@/lib/componentFactory/flowComponent";
 import { useObjectRef } from "@react-aria/utils";
-import { useManagedValue } from "@/lib/hooks/useManagedValue";
 import { type PropsContext, PropsContextProvider } from "@/lib/propsContext";
-import { TunnelProvider, TunnelExit } from "@mittwald/react-tunnel";
+import { TunnelExit, TunnelProvider } from "@mittwald/react-tunnel";
 import {
   modifyValueByMarkdownSyntax,
   scrollToCursor,
@@ -23,13 +16,13 @@ import {
   type InsertType,
   modifyValueByType,
 } from "@/components/MarkdownEditor/lib/modifyValueByType";
+import { useControlledHostValueProps } from "@/lib/remote/useControlledHostValueProps";
 
 export type MarkdownEditorMode = "editor" | "preview";
 
-export type MarkdownEditorProps = Omit<TextAreaProps, "ref"> &
-  Pick<MarkdownProps, "headingOffset" | "ref"> & {
-    inputRef?: RefObject<HTMLTextAreaElement | null>;
-  };
+export interface MarkdownEditorProps
+  extends TextAreaProps,
+    Pick<MarkdownProps, "headingOffset"> {}
 
 /** @flr-generate all */
 export const MarkdownEditor = flowComponent("MarkdownEditor", (props) => {
@@ -41,48 +34,14 @@ export const MarkdownEditor = flowComponent("MarkdownEditor", (props) => {
     rows = 5,
     autoResizeMaxRows,
     headingOffset,
+    value,
+    onChange,
     ref,
-    inputRef,
     ...rest
-  } = props;
+  } = useControlledHostValueProps(props);
 
-  const localRef = useObjectRef(ref);
-  const localInputRef = useObjectRef(inputRef);
-
+  const inputRef = useObjectRef(ref);
   const [mode, setMode] = useState<MarkdownEditorMode>("editor");
-  const { value, handleOnChange } = useManagedValue(props);
-
-  const selectionPresent = useRef<{
-    shouldScrollToCursor: boolean;
-    selectionStart: number | null;
-    selectionEnd: number | null;
-  } | null>(null);
-
-  useEffect(() => {
-    const inputRef = localInputRef.current;
-    const present = selectionPresent.current;
-
-    if (!present || !inputRef) {
-      return;
-    }
-
-    setTimeout(() => {
-      requestAnimationFrame(() => {
-        requestAnimationFrame(() => {
-          inputRef.setSelectionRange(
-            present.selectionStart,
-            present.selectionEnd,
-          );
-
-          if (present.shouldScrollToCursor) {
-            scrollToCursor(value, inputRef);
-          }
-        });
-      });
-
-      selectionPresent.current = null;
-    }, 0);
-  }, [selectionPresent.current, localInputRef.current, value]);
 
   const rootClassName = clsx(
     styles.markdownEditor,
@@ -95,40 +54,41 @@ export const MarkdownEditor = flowComponent("MarkdownEditor", (props) => {
       return;
     }
 
-    const modifyParams = modifyValueByMarkdownSyntax(value, localInputRef);
-
+    const modifyParams = modifyValueByMarkdownSyntax(value, inputRef);
     if (!modifyParams) {
       return;
     }
 
     const { newValue, newSelectionStart, newSelectionEnd } = modifyParams;
 
+    requestAnimationFrame(() => {
+      if (inputRef.current) {
+        inputRef.current.value = newValue;
+        inputRef.current?.setSelectionRange(newSelectionStart, newSelectionEnd);
+        scrollToCursor(newValue, inputRef.current);
+      }
+    });
+
     event.preventDefault();
-    localInputRef.current?.blur();
-
-    selectionPresent.current = {
-      shouldScrollToCursor: true,
-      selectionStart: newSelectionStart,
-      selectionEnd: newSelectionEnd,
-    };
-
-    handleOnChange(newValue);
+    onChange(newValue);
   };
 
   const handleToolButtonPressed = (type: InsertType) => {
     const { newValue, newSelectionStart, newSelectionEnd } = modifyValueByType(
       value,
       type,
-      localInputRef,
+      inputRef,
     );
 
-    selectionPresent.current = {
-      shouldScrollToCursor: false,
-      selectionStart: newSelectionStart,
-      selectionEnd: newSelectionEnd,
-    };
+    requestAnimationFrame(() => {
+      if (inputRef.current) {
+        inputRef.current.value = newValue;
+        inputRef.current.setSelectionRange(newSelectionStart, newSelectionEnd);
+        inputRef.current.focus();
+      }
+    });
 
-    handleOnChange(newValue);
+    onChange(newValue);
   };
 
   const propsContext: PropsContext = {
@@ -138,7 +98,7 @@ export const MarkdownEditor = flowComponent("MarkdownEditor", (props) => {
   };
 
   return (
-    <div ref={localRef} className={rootClassName}>
+    <div className={rootClassName}>
       <TunnelProvider>
         <TunnelExit id="label" />
         <Toolbar
@@ -152,18 +112,18 @@ export const MarkdownEditor = flowComponent("MarkdownEditor", (props) => {
           aria-hidden={mode === "preview"}
           isReadOnly={isReadOnly || mode === "preview"}
           isDisabled={isDisabled}
-          ref={localInputRef}
+          ref={inputRef}
           value={value}
           rows={rows}
           autoResizeMaxRows={autoResizeMaxRows}
-          onChange={handleOnChange}
+          onChange={onChange}
           onKeyDown={handleKeyDown}
         >
           <Markdown
             headingOffset={headingOffset}
             className={styles.markdown}
             style={{
-              height: localInputRef.current?.offsetHeight,
+              height: inputRef.current?.offsetHeight,
             }}
           >
             {value}