+++ packages/components/src/components/Overlay/Overlay.tsx
@@ -6,7 +6,6 @@ import { useOverlayController } from "@/lib/controller";
 import OverlayContextProvider from "@/lib/controller/overlay/OverlayContextProvider";
 import type { PropsWithClassName } from "@/lib/types/props";
 import OverlayContentView from "@/views/OverlayContentView";
-import { ClearPropsContext } from "@/index/default";
 
 export interface OverlayProps extends PropsWithChildren, PropsWithClassName {
   ref?: Ref<HTMLDivElement>;
@@ -41,19 +40,17 @@ export const Overlay: FC<OverlayProps> = (props) => {
   const rootClassName = clsx(styles.overlay, className);
 
   return (
-    <ClearPropsContext keep={overlayType}>
-      <OverlayContentView
-        onOpenChange={(isOpen) => controller.setOpen(isOpen)}
-        isOpen={isOpen}
-        ref={ref}
-        isDismissable={isDismissable}
-        className={rootClassName}
-      >
-        <OverlayContextProvider type="Modal" controller={controller}>
-          {isOpen && children}
-        </OverlayContextProvider>
-      </OverlayContentView>
-    </ClearPropsContext>
+    <OverlayContentView
+      onOpenChange={(isOpen) => controller.setOpen(isOpen)}
+      isOpen={isOpen}
+      ref={ref}
+      isDismissable={isDismissable}
+      className={rootClassName}
+    >
+      <OverlayContextProvider type="Modal" controller={controller}>
+        {isOpen && children}
+      </OverlayContextProvider>
+    </OverlayContentView>
   );
 };