+++ packages/components/src/lib/remote/eventHandlerContext.ts
@@ -1,0 +1,12 @@
+import { createCascade } from "context";
+
+export interface EventHandlerContext {
+  remoteEvent?: {
+    type: string;
+  };
+  [key: string]: unknown;
+}
+
+export const eventHandlerContext = createCascade<EventHandlerContext>();
+
+export const getRemoteEvent = () => eventHandlerContext.use()?.remoteEvent;